fix: Temporary 'fix' for https://github.com/rust-lang/rust/issues/56650.
The build system fails if there are spaces in the path leading to the
rust repository.  This is a quick fix that detects if there are any
spaces in the path leading to the rust repository, and if there are,
quits with a message warning the user about the problem.

Signed-off-by: Cem Karan <cem.f.karan.civ@army.mil>